Dynamic user group assignment
System plugin refactoring

gh-46
