+++ .idea/copyright/Akeeba.xml
@@ -1,0 +1,8 @@
+<component name="CopyrightManager">
+  <copyright>
+    <option name="allowReplaceRegexp" value="Copyright\s*\(c\)\s*2016-20[0-9]{2,2}\s*Nicholas K. Dionysopoulos" />
+    <option name="keyword" value="Copyright\s*\(c\)\s*2016-20[0-9]{2,2}\s*Nicholas K. Dionysopoulos" />
+    <option name="notice" value=" @package   AkeebaSocialLogin&#10; @copyright Copyright (c)2016-&amp;#36;today.year Nicholas K. Dionysopoulos / Akeeba Ltd&#10; @license   GNU General Public License version 3, or later" />
+    <option name="myName" value="Akeeba" />
+  </copyright>
+</component>
