+++ plugins/system/sociallogin/sociallogin.xml
@@ -15,6 +15,7 @@
 
 	<files>
 		<filename plugin="sociallogin">sociallogin.php</filename>
+		<folder>Features</folder>
 		<folder>fields</folder>
 		<folder>layout</folder>
 		<folder>Library</folder>
